refactor: move test in right folder + add it to index
