+++ packages/cmf/__tests__/Inject.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import Inject from './Inject.component';
+import Inject from '../src/Inject.component';
 
 describe('Inject', () => {
 	it('should render', () => {
