--- packages/cmf/src/Inject/index.js
@@ -1,3 +1,0 @@
-import Inject from './Inject.component';
-
-export default Inject;
